% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NaviCell_S4.R
\name{nvLaunchBrowser}
\alias{nvLaunchBrowser}
\title{launch the (default) client browser}
\usage{
nvLaunchBrowser(object)
}
\arguments{
\item{nv}{NaviCell object}
}
\description{
launch the (default) client browser
}

